+++ src/Kook.Net.Rest/API/Rest/CreateDirectMessageParams.cs
@@ -20,7 +20,7 @@ internal class CreateDirectMessageParams
 
     [JsonPropertyName("template_id")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? TemplateId { get; set; }
+    public ulong? TemplateId { get; set; }
 
     [JsonPropertyName("content")]
     public required string Content { get; set; }
