+++ src/Kook.Net.Rest/API/Rest/CreateMessageParams.cs
@@ -13,7 +13,7 @@ internal class CreateMessageParams
 
     [JsonPropertyName("template_id")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? TemplateId { get; set; }
+    public ulong? TemplateId { get; set; }
 
     [JsonPropertyName("content")]
     public required string Content { get; set; }
