+++ test/Kook.Net.Tests.Unit/MockedEntities/MockedVoiceChannel.cs
@@ -154,7 +154,7 @@ internal sealed class MockedVoiceChannel : IVoiceChannel
         throw new NotImplementedException();
 
     /// <inheritdoc />
-    public Task<Cacheable<IUserMessage, Guid>> SendTextAsync<T>(int templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
+    public Task<Cacheable<IUserMessage, Guid>> SendTextAsync<T>(ulong templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
         JsonSerializerOptions? jsonSerializerOptions = null, RequestOptions? options = null) =>
         throw new NotImplementedException();
 
@@ -169,7 +169,7 @@ internal sealed class MockedVoiceChannel : IVoiceChannel
         throw new NotImplementedException();
 
     /// <inheritdoc />
-    public Task<Cacheable<IUserMessage, Guid>> SendCardsAsync<T>(int templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
+    public Task<Cacheable<IUserMessage, Guid>> SendCardsAsync<T>(ulong templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
         JsonSerializerOptions? jsonSerializerOptions = null, RequestOptions? options = null) =>
         throw new NotImplementedException();
